server/pyproject.toml
[tool.poetry]
name = "klokanuv-kufr"
version = "0.1.0"
description = ""
authors = ["Jakub Zárybnický <jakub@zarybnicky.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
Faker = "^13.11.0"
django-allauth = "^0.50.0"
django-invitations = { git = "https://github.com/jazzband/django-invitations.git", rev = "045dc4d5" }

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']